decimal to binary

+++ DSA_Completer/Trailing_zero.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 using namespace std;
-int countZero(long n)
+int countZero(int n)
 {
     int fact = 1;
     for (int i = 2; i <= n; i++)
@@ -18,7 +18,7 @@ int countZero(long n)
 
 int main()
 {
-    long n;
+    int n;
     cout << "Enter the number n=";
     cin>>n;
     cout<<countZero(n);
